Added fly_to method in models.Player and made end_game script
Made a script which ends the game for all users
While making the script, I realized that there is code duplication in the script and the exit_game view. There will also be repetition in the fly_to view. So I made a fly_to method in the Player model.
I also made the exit_game view use the fly_to method
To test this method, I had to replace constants in config.json and questions.json